requirement.txt
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Classful==0.14.1
Flask-HTTPAuth==3.2.4
Flask-JWT-Extended==4.1.0
Flask-KVSession-fork==0.6.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-Restless==0.17.0
Flask-Security-Too==4.0.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
urllib3==1.26.2
passlib==1.7.4
python-dotenv-0.19.2